update assistant folder
