--- assistant/RagAdapter.py
@@ -1,5 +1,0 @@
-from flask_restful import Resource
-
-class RagAdapter(Resource):
-    def post(self):
-        return {"response": "hola"}
