+++ assistant/adapters/AnalyzePdfAdapter.py
@@ -1,0 +1,5 @@
+from flask_restful import Resource
+
+class AnalyzePdfAdapter(Resource):
+    def post(self):
+        return {"response": "hola desde pdf adapter"}
