+++ assistant/adapters/RagAdapter.py
@@ -1,0 +1,5 @@
+from flask_restful import Resource
+
+class RagAdapter(Resource):
+    def post(self):
+        return {"response": "hola desde RAG"}
